auto_applier_v2/requirements.txt
playwright
playwright-stealth
customtkinter
sqlalchemy
openpyxl
requests
PyPDF2
pypdf
pandas
fastapi
uvicorn
authlib
PyJWT
openai
httpx
